backend/requirements.txt
Django==4.2.25
djangorestframework==3.14.0
pandas==2.1.4
openpyxl==3.1.2
python-magic==0.4.27
psycopg2-binary==2.9.9
pytest==8.4.2
pytest-django==4.11.1
pytest-cov==7.0.0
